requirements.txt
pyaudio
google-genai
Pillow
requests
beautifulsoup4
playwright
pyautogui
pyperclip
opencv-python
numpy
mss
psutil
send2trash
comtypes; sys_platform == "win32"
pycaw; sys_platform == "win32"
win10toast; sys_platform == "win32"
duckduckgo-search
youtube-transcript-api
python-socks[asyncio]
httpx[socks]
websockets>=14.0